requirements.txt
Django==4.0.5
pymongo==3.12.3
djongo==1.3.6
sqlparse==0.2.4
pytz==2025.2
#selenium==4.15.2
gunicorn==21.2.0
#chromedriver-py==126.0.6478.126
pyvirtualdisplay==3.0
#python-telegram-bot==13.13
pyautogui==0.9.54
telebot==0.0.5
legacy-cgi==2.6.2
pyTelegramBotAPI==4.27.0
djangorestframework-simplejwt
django-cors-headers